requirements.txt
arrow==1.3.0
asgiref==3.10.0
binaryornot==0.4.4
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cookiecutter==2.6.0
cryptography==46.0.2
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.12.0
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
dotenv==0.9.9
filelock==3.20.0
fsspec==2025.9.0
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
hf-xet==1.1.10
honcho==2.0.0
httplib2==0.31.0
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
